refactor(refs T30969): move DpDraggable to components root
