+++ src/components/core/index.js
@@ -7,7 +7,6 @@ import DpCopyPasteButton from './DpCopyPasteButton'
 import DpDashboardTaskCard from './DpDashboardTaskCard'
 import DpDataTable from './DpDataTable/DpDataTable'
 import DpDataTableExtended from './DpDataTable/DpDataTableExtended'
-import DpDraggable from "./DpDraggable"
 import DpEditableList from './DpEditableList'
 import DpEditor from './DpEditor/DpEditor'
 import DpNotifyContainer from './notify/DpNotifyContainer'
@@ -36,7 +35,6 @@ export {
   DpDashboardTaskCard,
   DpDataTable,
   DpDataTableExtended,
-  DpDraggable,
   DpEditableList,
   DpEditor,
   DpNotifyContainer,
